checking the online commit messages
It's about living dreamed life

The plot of "Cinderella" is that a young girl (who is treated terribly by her stepsisters),
experiences a bit of magic from her fairy godmother and is able to go to a ball dressed as a beautiful princess.
There, she falls in love with a prince.
The following day, she goes back but leaves her slipper.